cv_pipeline/requirements.txt
# Core ML/DL libraries
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# MLflow для трекинга экспериментов
mlflow>=2.8.0

# Image processing
Pillow>=10.0.0
imagehash>=4.3.1
opencv-python>=4.8.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Logging and monitoring
loguru>=0.7.0

